@@ -1,7 +1,7 @@
 /* eslint-disable import/max-dependencies */
 /* eslint-disable no-console */
 import React from 'react';
-import ReactDOM from 'react-dom';
+import { createRoot } from 'react-dom/client';
 
 // eslint-disable-next-line import/no-unresolved
 import { StatefulTabs, Pagination, ToastsContainer } from '../dist/main';
@@ -18,7 +18,9 @@ import { UncontrolledFormExamples } from './UncontrolledFormExamples';
 import { ToastsWithMessageFormatterExamples } from './ToastsWithMessageFormatterExamples';
 import { ToastWithCustomState } from './ToastWithCustomState';
 
-ReactDOM.render(
+const root = createRoot(document.getElementById('root'));
+
+root.render(
   <div className="mt-3">
     <React.StrictMode>
       <StatefulTabs
@@ -85,8 +87,7 @@ ReactDOM.render(
         ]}
       />
     </React.StrictMode>
-  </div>,
-  document.getElementById('root')
+  </div>
 );
 
 // eslint-disable-next-line no-undef
