@@ -1,6 +1,6 @@
 'use strict';
 
-const webpack = require('webpack');
+const ReactRefreshWebpackPlugin = require('@pmmmwh/react-refresh-webpack-plugin');
 
 module.exports = [
   {
@@ -28,14 +28,35 @@ module.exports = [
   {
     entry: './demo/demo.jsx',
     output: {
-      path: `${__dirname}/demo`,
       filename: 'demo.js',
+      path: `${__dirname}/demo`,
+      publicPath: '/demo/',
     },
     devServer: {
-      contentBase: `${__dirname}/demo`,
+      static: {
+        directory: `${__dirname}/demo`,
+      },
       hot: true,
+      port: 8080,
     },
-    plugins: [new webpack.HotModuleReplacementPlugin()],
+    module: {
+      rules: [
+        {
+          test: /\.(js|jsx)$/,
+          exclude: /node_modules/,
+          use: {
+            loader: 'babel-loader',
+            options: {
+              plugins: [require.resolve('react-refresh/babel')],
+            },
+          },
+        },
+      ],
+    },
+    resolve: {
+      extensions: ['*', '.js', '.jsx'],
+    },
+    plugins: [new ReactRefreshWebpackPlugin()],
   },
   {
     entry: './docs/docs.jsx',
@@ -45,7 +66,6 @@ module.exports = [
     },
   },
 ].map((setup) => ({
-  ...setup,
   module: {
     rules: [
       {
@@ -60,4 +80,5 @@ module.exports = [
   resolve: {
     extensions: ['*', '.js', '.jsx'],
   },
+  ...setup,
 }));
